sd_training_project/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
Pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0
huggingface-hub>=0.15.0 